requirements.txt
numpy
pytorch
opencv-python
pandas
Pillow